+++ R/sentopics.R
@@ -41,7 +41,7 @@ NULL
 #' @format A [quanteda::corpus] object.
 #' @seealso [ECB_press_conferences_tokens]
 #'
-#' @source \url{https://www.ecb.europa.eu/press/key/date/html/index.en.html}.
+#' @source \url{https://www.ecb.europa.eu/press/pubbydate/html/index.en.html}.
 "ECB_press_conferences"
 
 #' Tokenized press conferences
@@ -63,7 +63,7 @@ NULL
 #' @format A [quanteda::tokens] object.
 #' @seealso [ECB_press_conferences]
 #'
-#' @source \url{https://www.ecb.europa.eu/press/key/date/html/index.en.html}.
+#' @source \url{https://www.ecb.europa.eu/press/pubbydate/html/index.en.html}.
 "ECB_press_conferences_tokens"
 
 #' Loughran-McDonald lexicon